Remove unavailable item from active sync requests
When got an `item_not_available_message` from a peer, the item was
removed from the `sync_items_requested_from_peer` list of the peer,
but it was left in the `_active_sync_requests` list.
When disconnecting from the peer, since the
`sync_items_requested_from_peer` list of the peer no longer contains
the item, the cleanup process does not remove it from the
`_active_sync_requests` list either.
As a result, the node will not request the item from another peer,
thus gets stuck.